[Chore] criteria Dict로 변경
