+++ PhysicalTrack/PhysicalTrack/Domain/PushUpRecord.swift
@@ -24,7 +24,7 @@ struct PushUpRecord: Equatable {
     
     init(for grade: Grade) {
         self.targetSeconds = 10
-        self.targetCount = PushUp.getValue(grade).upperBound
+        self.targetCount = grade == .elite ? PushUp.criteriaDict[grade]!.lowerBound : PushUp.criteriaDict[grade]!.upperBound
         self.time = 0
         self.count = 0
     }
