+++ PhysicalTrack/PhysicalTrack/Domain/Workout.swift
@@ -7,40 +7,22 @@
 
 import Foundation
 
-struct GradeCriteria<W: Workout>: Equatable, Hashable {
-    let grade: Grade
-    var value: ClosedRange<Int> {
-        W.getValue(grade)
-    }
-    
-    var description: String {
-        guard value.lowerBound != Int.min 
-        else { return String(value.upperBound) + " 이하" }
-        
-        guard value.upperBound != Int.max
-        else { return String(value.lowerBound) + " 이상" }
-        
-        return String(value.lowerBound) + " - " + String(value.upperBound)
-    }
-}
-
 protocol Workout: Equatable {
-    static func getValue(_ grade: Grade) -> ClosedRange<Int>
-    
     static func judgeGrade(_ value: Int) -> Grade
     
-    static var list: [GradeCriteria<Self>] { get }
+    static var criteriaDict: [Grade: ClosedRange<Int>] { get }
+    static var criteriaArray: [GradeCriteria<Self>] { get }
 }
 
 extension Workout {
-    static var list: [GradeCriteria<Self>] {
+    static var criteriaArray: [GradeCriteria<Self>] {
         return Grade.allCases.map { .init(grade: $0) }
     }
     
     static func judgeGrade(_ value: Int) -> Grade {
         
         for grade in Grade.allCases {
-            if Self.getValue(grade).contains(value) { return grade }
+            if Self.criteriaDict[grade]!.contains(value) { return grade }
         }
         return .failed
     }
